pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["bnm <bnm@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
setuptools = "^62.1.0"
grpcio = "^1.47.0"
grpcio-tools = "^1.47.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.4"
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.8.0"
flake8-isort = "^4.1.1"
flake8-pytest-style = "^1.6.0"
pytest = "^7.1.1"
toml = "^0.10.2"
flake8-pytest = "^1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.bandit]
exclude_dirs = [".venv/", ".git/"]
assert_used = { skips = ["*/test_*.py", "*/test_*.py", "*/conftest.py"] }
